Feat: 컴포넌트 합침
- 글 버튼, 필터 , 썸네일 컴포넌트 병합
