--- client/src/component/selectBtn/SelectBtn.css
@@ -1,36 +1,0 @@
-.selectBtnBackgroundSize{
-    width: 100%;
-    height: 50px;
-    background-color: #FFFFFF;
-    padding-bottom: 50px;
-}
-
-.selectBtnSize{
-    flex-wrap: nowrap;
-    display: flex;
-    max-width: 1000px;
-    padding-top: 15px;
-    padding-left: 80px;
-    padding-right: 80px;
-    text-align: left;
-    width: 67%;
-    margin-left: 9%;
-}
-
-.selectBtn{
-    white-space:nowrap;
-    display: flex;
-    /* padding-left: 10px;
-    padding-right: 10px; */
-    padding: 10px;
-    background-color: rgba(0,0,0,0);
-    border: 0;
-    font-size: 18px;
-    font-weight: bold;
-    color: #7c7c7c;
-}
-
-.selectBtn:hover{
-    color:black;
-}
-
